Fix calling existing ctor with MethodInvoker; share tests with invokers
