+++ src/libraries/System.Private.CoreLib/src/System/Reflection/MethodInvoker.cs
@@ -66,7 +66,12 @@ namespace System.Reflection
             {
                 // This is useful for calling a constructor on an already-initialized object
                 // such as created from RuntimeHelpers.GetUninitializedObject(Type).
-                return new MethodInvoker(rci);
+                MethodInvoker invoker = new MethodInvoker(rci);
+
+                // Use the interpreted version to avoid having to generate a new method that doesn't allocate.
+                invoker._strategy = GetStrategyForUsingInterpreted();
+
+                return invoker;
             }
 
             throw new ArgumentException(SR.Argument_MustBeRuntimeMethod, nameof(method));
@@ -181,7 +186,7 @@ namespace System.Reflection
 
         private object? InvokeImpl(object? obj, object? arg1, object? arg2, object? arg3, object? arg4)
         {
-            if ((_invocationFlags & (InvocationFlags.NoInvoke | InvocationFlags.ContainsStackPointers)) != 0)
+            if ((_invocationFlags & (InvocationFlags.NoInvoke | InvocationFlags.ContainsStackPointers | InvocationFlags.NoConstructorInvoke)) != 0)
             {
                 ThrowForBadInvocationFlags();
             }
