+++ src/libraries/System.Reflection/tests/ConstructorInvokerTests.cs
@@ -1,13 +1,26 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Runtime.CompilerServices;
 using Xunit;
 
 namespace System.Reflection.Tests
 {
-    public class ConstructorInvokerTests
+    /// <summary>
+    /// These tests use the shared tests from the base class with ConstructorInvoker.Invoke.
+    /// </summary>
+    public sealed class ConstructorInvokerTests : ConstructorCommonTests
     {
+        public override object Invoke(ConstructorInfo constructorInfo, object?[]? parameters)
+        {
+            return ConstructorInvoker.Create(constructorInfo).Invoke(new Span<object>(parameters));
+        }
+
+        public override object? Invoke(ConstructorInfo constructorInfo, object obj, object?[]? parameters)
+        {
+            return MethodInvoker.Create(constructorInfo).Invoke(obj, new Span<object>(parameters));
+        }
+
+        protected override bool IsExceptionWrapped => false;
 
         [Fact]
         public void Args_0()
@@ -162,16 +175,13 @@ namespace System.Reflection.Tests
         }
 
         [Fact]
-        public void ExistingInstance()
+        public void Invoke_StaticConstructor_NullObject_NullParameters()
         {
-            ConstructorInfo ci = typeof(TestClass).GetConstructor(BindingFlags.Public | BindingFlags.Instance, Type.EmptyTypes);
-            TestClass tc = (TestClass)RuntimeHelpers.GetUninitializedObject(typeof(TestClass));
-            Assert.Null(tc._args);
+            ConstructorInfo[] constructors = GetConstructors(typeof(ClassWithStaticConstructor));
+            Assert.Equal(1, constructors.Length);
 
-            MethodInvoker invoker = MethodInvoker.Create(ci);
-            object? obj = invoker.Invoke(tc);
-            Assert.Equal("0", tc._args);
-            Assert.Null(obj);
+            // Invoker classes do not support calling class constructors; use standard reflection for that.
+            Assert.Throws<MemberAccessException>(() => Invoke(constructors[0], null, new object[] { }));
         }
 
         private class TestClass
