+++ src/libraries/System.Reflection/tests/MethodInvokerTests.cs
@@ -2,13 +2,22 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections.Generic;
-using System.Linq;
 using Xunit;
 
 namespace System.Reflection.Tests
 {
-    public class MethodInvokerTests
+    /// <summary>
+    /// These tests use the shared tests from the base class with MethodInvoker.Invoke.
+    /// </summary>
+    public class MethodInvokerTests : MethodCommonTests
     {
+        public override object? Invoke(MethodInfo methodInfo, object? obj, object?[]? parameters)
+        {
+            return MethodInvoker.Create(methodInfo).Invoke(obj, new Span<object>(parameters));
+        }
+
+        protected override bool SupportsMissing => false;
+
         [Fact]
         public void NullTypeValidation()
         {
@@ -284,11 +293,6 @@ namespace System.Reflection.Tests
         {
             MethodInvoker invoker = MethodInvoker.Create(typeof(TestClass).GetMethod(nameof(TestClass.VerifyThisObj)));
             Assert.Throws<TargetException>(() => invoker.Invoke(obj: null));
-        }
-
-        private static MethodInfo GetMethod(Type type, string name)
-        {
-            return type.GetTypeInfo().GetMethods(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.Instance).First(method => method.Name.Equals(name));
         }
 
         public static IEnumerable<object[]> Invoke_TestData() => MethodInfoTests.Invoke_TestData();
