+++ src/libraries/System.Reflection/tests/System.Reflection.Tests.csproj
@@ -14,14 +14,12 @@
     <RdXmlFile Include="default.rd.xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="$(CommonTestPath)System\Reflection\MockParameterInfo.cs"
-             Link="Common\System\Reflection\MockParameterInfo.cs" />
-    <Compile Include="$(CommonTestPath)System\MockType.cs"
-             Link="Common\System\MockType.cs" />
-    <Compile Include="$(CommonTestPath)System\IO\TempFile.cs"
-             Link="Common\System\IO\TempFile.cs" />
+    <Compile Include="$(CommonTestPath)System\Reflection\MockParameterInfo.cs" Link="Common\System\Reflection\MockParameterInfo.cs" />
+    <Compile Include="$(CommonTestPath)System\MockType.cs" Link="Common\System\MockType.cs" />
+    <Compile Include="$(CommonTestPath)System\IO\TempFile.cs" Link="Common\System\IO\TempFile.cs" />
     <Compile Include="AssemblyNameTests.cs" />
     <Compile Include="AssemblyTests.cs" />
+    <Compile Include="ConstructorCommonTests.cs" />
     <Compile Include="ConstructorInfoTests.cs" />
     <Compile Include="ConstructorInvokerTests.cs" />
     <Compile Include="CustomAttributeTests.cs" />
@@ -31,6 +29,7 @@
     <Compile Include="GetTypeTests.cs" />
     <Compile Include="ManifestResourceInfoTests.cs" />
     <Compile Include="MemberInfoTests.cs" />
+    <Compile Include="MethodCommonTests.cs" />
     <Compile Include="MethodInfoTests.cs" />
     <Compile Include="MethodInvokerTests.cs" />
     <Compile Include="ModuleTests.cs" />
@@ -42,8 +41,7 @@
     <Compile Include="ExceptionTests.cs" />
     <Compile Include="PointerTests.cs" />
     <Compile Include="TypeTests.Constraints.cs" />
-    <Compile Include="TestAssembly\EquivalentValueType.cs"
-             Link="EquivalentValueType.cs" />
+    <Compile Include="TestAssembly\EquivalentValueType.cs" Link="EquivalentValueType.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resources\EmbeddedImage.png">
